[Core] Extract mailer sender from ShopBundle
