+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/command_handlers.xml
@@ -78,8 +78,8 @@
         <service id="Sylius\Bundle\ApiBundle\CommandHandler\Checkout\CompleteOrderHandler">
             <argument type="service" id="sylius.repository.order"/>
             <argument type="service" id="sm.factory" />
-            <argument type="service" id="sylius.email_manager.order" />
-            <tag name="messenger.message_handler" bus="sylius_default.bus"/>
+            <argument type="service" id="sylius.mailer.order_email_manager" />
+            <tag name="messenger.message_handler" />
         </service>
 
         <service id="Sylius\Bundle\ApiBundle\CommandHandler\Cart\ChangeItemQuantityInCartHandler">
