+++ src/Sylius/Bundle/CoreBundle/Mailer/OrderEmailManager.php
@@ -11,9 +11,9 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\EmailManager;
+namespace Sylius\Bundle\CoreBundle\Mailer;
 
-use Sylius\Bundle\CoreBundle\Mailer\Emails;
+use Sylius\Component\Core\Model\CustomerInterface;
 use Sylius\Component\Core\Model\OrderInterface;
 use Sylius\Component\Mailer\Sender\SenderInterface;
 
@@ -29,14 +29,17 @@ final class OrderEmailManager implements OrderEmailManagerInterface
 
     public function sendConfirmationEmail(OrderInterface $order): void
     {
+        /** @var CustomerInterface $customer */
+        $customer = $order->getCustomer();
+
         $this->emailSender->send(
             Emails::ORDER_CONFIRMATION,
-            [$order->getCustomer()->getEmail()],
+            [$customer->getEmail()],
             [
                 'order' => $order,
                 'channel' => $order->getChannel(),
                 'localeCode' => $order->getLocaleCode(),
-            ])
-        ;
+            ]
+        );
     }
 }
