+++ src/Sylius/Bundle/CoreBundle/spec/Mailer/OrderEmailManagerSpec.php
@@ -11,11 +11,10 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Bundle\ShopBundle\EmailManager;
+namespace spec\Sylius\Bundle\CoreBundle\Mailer;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Bundle\CoreBundle\Mailer\OrderEmailManagerInterface as CoreOrderEmailManagerInterface;
-use Sylius\Bundle\ShopBundle\EmailManager\OrderEmailManagerInterface;
+use Sylius\Bundle\CoreBundle\Mailer\OrderEmailManagerInterface;
 use Sylius\Component\Core\Model\ChannelInterface;
 use Sylius\Component\Core\Model\CustomerInterface;
 use Sylius\Component\Core\Model\OrderInterface;
@@ -23,9 +22,9 @@ use Sylius\Component\Mailer\Sender\SenderInterface;
 
 final class OrderEmailManagerSpec extends ObjectBehavior
 {
-    function let(SenderInterface $sender, CoreOrderEmailManagerInterface $decoratedEmailManager): void
+    function let(SenderInterface $sender): void
     {
-        $this->beConstructedWith($sender, $decoratedEmailManager);
+        $this->beConstructedWith($sender);
     }
 
     function it_implements_an_order_email_manager_interface(): void
@@ -33,23 +32,12 @@ final class OrderEmailManagerSpec extends ObjectBehavior
         $this->shouldImplement(OrderEmailManagerInterface::class);
     }
 
-    function it_delegates_email_sending_to_core_implementation(
-        CoreOrderEmailManagerInterface $decoratedEmailManager,
-        OrderInterface $order
-    ): void {
-        $decoratedEmailManager->sendConfirmationEmail($order)->shouldBeCalled();
-
-        $this->sendConfirmationEmail($order);
-    }
-
-    function it_sends_an_order_confirmation_email_if_core_email_manager_is_not_set(
+    function it_sends_an_order_confirmation_email(
         SenderInterface $sender,
         OrderInterface $order,
         ChannelInterface $channel,
         CustomerInterface $customer
     ): void {
-        $this->beConstructedWith($sender, null);
-
         $order->getChannel()->willReturn($channel);
         $order->getLocaleCode()->willReturn('en_US');
         $order->getCustomer()->willReturn($customer);
